- added cart

+++ Omadiko.Entities/Models/Membership.cs
@@ -10,10 +10,9 @@ namespace Omadiko.Entities.Models
     {
         public int MembershipId { get; set; }
         public string Type { get; set; }
-        public float SignUpFee { get; set; }
+        public int SignUpFee { get; set; }
         public byte DurationInMonths { get; set; }
         public byte DiscountRate { get; set; }
-
 
     }
 }
